Implemented LLM calculation of calories, protein, fat & carbs

+++ app/core/db/tables.py
@@ -1,5 +1,5 @@
 from sqlalchemy import MetaData
-from sqlalchemy import Table, Column, Integer, TIMESTAMP, ForeignKey, String, Float
+from sqlalchemy import Table, Column, Integer, TIMESTAMP, ForeignKey, String, Float, Numeric
 from sqlalchemy.dialects.postgresql import JSONB
 
 metadata_obj = MetaData()
@@ -19,10 +19,10 @@ meal_entry = Table(
     Column("tg_user_id", Integer, ForeignKey("tg_user.tg_user_id"), nullable=False),
     Column("text", String(300), nullable=False),
     Column("created_at", TIMESTAMP),
-    Column("calories", Integer),
-    Column("protein", Integer),
-    Column("fat", Integer),
-    Column("carbs", Integer),
+    Column("calories", Numeric(6, 1)),
+    Column("protein", Numeric(6, 1)),
+    Column("fat", Numeric(6, 1)),
+    Column("carbs", Numeric(6, 1)),
     Column("llm_raw", JSONB),
-    Column("confidence", Float, nullable=False),
+    Column("confidence", Integer, nullable=False),
 )
